bpf: pass Geneve-DSR packet to host-stack
The purpose of this commit is to fix an issue where Geneve-DSR is broken when BPF-Masquerade is disabled.
https://github.com/cilium/cilium/issues/32189

The cause of this issue is that reply packets of Geneve-DSR were being masqueraded due to iptables rules.

The code added in this commit ensures that when Geneve-DSR packets are received, cil_from_overlay() passes them to the host-stack.
This causes the conntrack-entry to be inserted, preventing the packets from being masqueraded.

Signed-off-by: yushoyamaguchi <ysh.824@outlook.jp>