Fix ICMP header's checksum when SNAT RevNAT
- ExternalTrafficPolicy = cluster
- On the LB node, SNAT RevNAT is applied to ICMP error packets, which is reply from LB's backend-pod.
In such cases, we observed that the L4 checksum(checksum of ICMP-Header) in the ICMP error packet was incorrect. This commit fixes that bug.
https://github.com/cilium/cilium/issues/40827

The root cause was that during RevNAT, address/port rewriting was also applied to the original packet contained in the ICMP error payload, but the change was not reflected in the checksum of ICMP-Header.
Therefore, this commit implements the logic to recalculate the checksum accordingly and write the new checksum back into the packet.

Signed-off-by: Yusho Yamaguchi <ysh.824@outlook.jp>