port replace before l4 checksum replace
When rewriting the inner header within the payload of an ICMP error packet, the l4_modify_port() function is also used.
In such cases, the inner L4 header may be truncated and not include the checksum field.

When that happens, the function returns DROP_CSUM_L4, but the caller handles this error and continues processing.
https://github.com/cilium/cilium/pull/40194/commits/30a6dd2158aea24b9fbdd61438daefef2cf2786d

Before returning DROP_CSUM_L4, I want to modify the port in the inner L4 header first.

Signed-off-by: YushoYamaguchi <ysh.824@outlook.jp>